+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 public interface IDamageable
 {
-    void TakeDamage(int amount, bool isDead = false);
+    void TakeDamage(float amount, bool isDead = false);
     void Die(float destroyDelay);
 }
